Update model definitions
- Fix bug about createsuperuser command
- Remove year app and migrate into course app
