+++ calyx/src/calyx/settings.py
@@ -34,7 +34,6 @@ INSTALLED_APPS = [
     'rest_framework_jwt',
     'rest_framework_swagger',
     'djoser',
-    'years',
     'courses',
     'users',
     'labs'
