--- calyx/src/years/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from .models import Year
-
-admin.site.register(Year)
