--- calyx/src/years/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class YearsConfig(AppConfig):
-    name = 'years'
